+++ src/Sftp/Sftp/CHANGELOG.md
@@ -19,7 +19,7 @@
 -->
 
 ## Upcoming Release
-* Added `BufferSizeBytes` parameter to `Connect-AzSftp` cmdlet
+* Added `BufferSize` parameter to `Connect-AzSftp` cmdlet
     - Allows users to specify buffer size in bytes for SFTP file transfers using the sftp -B flag
     - Default value is 262144 (256 KB)
 * Added `StorageAccountEndpoint` parameter to `Connect-AzSftp` cmdlet